+++ src/diffusers/models/uvit_2d.py
@@ -21,6 +21,7 @@ from torch import nn
 from torch.utils.checkpoint import checkpoint
 
 from ..configuration_utils import ConfigMixin, register_to_config
+from ..loaders import PeftAdapterMixin
 from .attention import BasicTransformerBlock, SkipFFTransformerBlock
 from .attention_processor import (
     ADDED_KV_ATTENTION_PROCESSORS,
@@ -35,7 +36,7 @@ from .normalization import GlobalResponseNorm, RMSNorm
 from .resnet import Downsample2D, Upsample2D
 
 
-class UVit2DModel(ModelMixin, ConfigMixin):
+class UVit2DModel(ModelMixin, ConfigMixin, PeftAdapterMixin):
     _supports_gradient_checkpointing = True
 
     @register_to_config
